+++ src/limiter_ui.cpp
@@ -80,5 +80,5 @@ void LimiterUi::reset() {
 void LimiterUi::on_new_attenuation(double value) {
   attenuation->set_value(1.0 - value);
 
-  attenuation_label->set_text(level_to_str(util::linear_to_db(static_cast<float>(value)), 0));
+  attenuation_label->set_text(level_to_str(util::linear_to_db(value), 0));
 }
